requirements.txt
Authlib==1.3.0
python-dotenv==1.0.1
